+++ src/modules/logseq/types/settings.ts
@@ -36,8 +36,8 @@ export type LogSeqSettings = {
   maxEmbeddedDocuments: number
   includeURLScrapper: boolean
   openRouterAPIKey: string
-  openRouterModel: OpenRouterModelEnum
-  openRouterEmbeddingModel: OpenRouterModelEnum
+  openRouterModel: string
+  openRouterEmbeddingModel: string
   claudeAPIKey: string
   claudeModel: ClaudeModelEnum
   mistralAPIKey: string